refactor multiple duplicates of download code into a composable
