+++ frontend/src/pages/application/Snapshots/components/dialogs/SnapshotExportDialog.vue
@@ -34,6 +34,7 @@ import { ClipboardCopyIcon, RefreshIcon } from '@heroicons/vue/outline'
 import snapshotsApi from '../../../../../api/snapshots.js'
 
 import FormRow from '../../../../../components/FormRow.vue'
+import { downloadData } from '../../../../../composables/Download.js'
 import clipboardMixin from '../../../../../mixins/Clipboard.js'
 import alerts from '../../../../../services/alerts.js'
 
@@ -99,7 +100,7 @@ export default {
                     return data
                 }).then(data => {
                     const snapshotDate = this.snapshot.updatedAt.replace(/[-:]/g, '').replace(/\..*$/, '').replace('T', '-')
-                    this.download(data, `snapshot-${this.snapshot.id}-${snapshotDate}.json`)
+                    downloadData(data, `snapshot-${this.snapshot.id}-${snapshotDate}.json`)
                     alerts.emit('Snapshot exported.', 'confirmation')
                     this.$refs.dialog.close()
                 }).catch(err => {
@@ -108,22 +109,6 @@ export default {
                 }).finally(() => {
                     this.submitted = false
                 })
-            }
-        },
-        async download (data, name) {
-            const dataString = typeof data === 'string' ? data : JSON.stringify(data, null, 2)
-            const element = document.createElement('a')
-            try {
-                element.setAttribute('href', 'data:text/plain;charset=utf-8,' + encodeURIComponent(dataString))
-                element.setAttribute('download', name)
-                element.style.display = 'none'
-                document.body.appendChild(element)
-                element.click()
-            } catch (err) {
-                console.error(err)
-                throw err
-            } finally {
-                document.body.removeChild(element)
             }
         },
         generateRandomKey (length = 16) {
