+++ frontend/src/pages/instance/Snapshots/index.vue
@@ -74,6 +74,7 @@ import SectionTopMenu from '../../../components/SectionTopMenu.vue'
 import AssetDetailDialog from '../../../components/dialogs/AssetDetailDialog.vue'
 import SnapshotImportDialog from '../../../components/dialogs/SnapshotImportDialog.vue'
 import UserCell from '../../../components/tables/cells/UserCell.vue'
+import { downloadData } from '../../../composables/Download.js'
 import permissionsMixin from '../../../mixins/Permissions.js'
 import Alerts from '../../../services/alerts.js'
 import Dialog from '../../../services/dialog.js'
@@ -254,13 +255,7 @@ export default {
                 version: '0.0.0-' + snapshot.id,
                 dependencies: ss.settings?.modules || {}
             }
-            const element = document.createElement('a')
-            element.setAttribute('href', 'data:text/plain;charset=utf-8,' + encodeURIComponent(JSON.stringify(packageJSON, null, 2)))
-            element.setAttribute('download', 'package.json')
-            element.style.display = 'none'
-            document.body.appendChild(element)
-            element.click()
-            document.body.removeChild(element)
+            downloadData(packageJSON, 'package.json')
         },
         showDownloadSnapshotDialog (snapshot) {
             this.$refs.snapshotExportDialog.show(snapshot)
