+++ frontend/src/pages/team/Devices/dialogs/ProvisioningCredentialsDialog.vue
@@ -23,6 +23,7 @@
 import { DocumentDownloadIcon } from '@heroicons/vue/outline'
 import { mapState } from 'vuex'
 
+import { downloadData } from '../../../../composables/Download.js'
 import clipboardMixin from '../../../../mixins/Clipboard.js'
 import Alerts from '../../../../services/alerts.js'
 
@@ -40,13 +41,7 @@ export default {
     },
     methods: {
         downloadCredentials () {
-            const element = document.createElement('a')
-            element.setAttribute('href', 'data:text/plain;charset=utf-8,' + encodeURIComponent(this.credentials))
-            element.setAttribute('download', 'device.yml')
-            element.style.display = 'none'
-            document.body.appendChild(element)
-            element.click()
-            document.body.removeChild(element)
+            downloadData(this.credentials, 'device.yml')
         },
         close () {
             this.$refs.dialog.close()
